+++ src/app-lite.js
@@ -1,24 +1,12 @@
 import { LitElement, html, css } from 'lit'
 import page from 'page'
 
-/* material design modules */
-import '@material/mwc-drawer'
-import '@material/mwc-tab'
-import '@material/mwc-tab-bar'
-import '@material/mwc-top-app-bar'
-import '@material/mwc-fab'
-import '@material/mwc-icon'
-import '@material/mwc-icon-button'
-import '@material/mwc-list'
-import '@material/mwc-list/mwc-list-item'
-import '@material/mwc-linear-progress'
-import '@material/mwc-snackbar'
-import '@material/mwc-textfield'
-
 /* components */
 import { PendingContainer } from './components/pending-container'
 import { views } from './components/views'
 import { lazyLoad } from './components/lazy-load'
+
+import '@cicciosgamino/snack-bar'
 
 /* shared styles */
 import { sharedStyles } from './styles/shared-styles.js'
@@ -77,7 +65,6 @@ class AppLite extends PendingContainer(LitElement) {
     window.removeEventListener('resize', this._drawerOrTabsLayout)
     window.removeEventListener('online', this._goingOnline)
     window.removeEventListener('offline', this._goingOffline)
-
     super.disconnectedCallback()
   }
 
@@ -85,54 +72,15 @@ class AppLite extends PendingContainer(LitElement) {
     return [
       sharedStyles,
       css`
-
       [hidden] { 
         display: none !important; 
       }
-
       :host {
         display: block;
-        background: linear-gradient(
-          to bottom right, var(--mdc-theme-background), var(--my-color));
-
         /* drawer 100% full screen, default is 256px */ 
         --mdc-drawer-width: 100%;
       }
-
       /* All elements interested have online / offline class */
-
-
-      /* drawer */ 
-      .drawer-content {
-        background-color: var(--mdc-theme-background);
-      }
-
-      .drawer-close-icon {
-        display: block;
-        position: fixed;
-        background-color: var(--mdc-theme-primary);
-        /* bottom centered */ 
-        bottom: 2em;
-        left: 50%;
-        transform: translate(-50%, 0);
-      }
-
-      #container {
-        background-color: var(--mdc-theme-background);
-        /* overflow: hidden; */
-      }
-      
-      .main-content {
-        min-height: 300px;
-        padding: 48px 18px 0 18px;
-      }
-
-      /* fab */ 
-      #fab-menu {
-        position: fixed;
-        bottom: 1.5em;
-        right: 2em;
-      }
       `
     ]
   }
@@ -173,19 +121,17 @@ class AppLite extends PendingContainer(LitElement) {
   // handle back online
   _goingOnline () {
     this.offline = false
-    // TODO
-    const snack = this.shadowRoot.querySelector('mwc-snackbar')
-    snack.setAttribute('labelText', 'Online')
-    snack.show()
+    const snack = this.shadowRoot.querySelector('snack-bar')
+    snack.setAttribute('title', 'Online')
+    snack.setAttribute('active', '')
   }
 
   // handle going Offline
   _goingOffline () {
     this.offline = true
-    // TODO
-    const snack = this.shadowRoot.querySelector('mwc-snackbar')
-    snack.setAttribute('labelText', 'Offline')
-    snack.show()
+    const snack = this.shadowRoot.querySelector('snack-bar')
+    snack.setAttribute('title', 'Offline')
+    snack.setAttribute('active', '')
   }
 
   // handle the drawer or tabs layout to render base on screen
@@ -249,115 +195,26 @@ class AppLite extends PendingContainer(LitElement) {
 
   render () {
     const drawerLayout = html`
-      <mwc-drawer
-            hasHeader 
-            type="modal" 
-            ?open="${this.drawerIsOpen}" 
-            @MDCDrawer:opened="${() => this.drawerIsOpen = true}"
-            @MDCDrawer:closed="${() => this.drawerIsOpen = false}">
-
-            <span slot="title">Drawer Title</span>
-            <span slot="subtitle">subtitle</span>
-
-            <!-- Drawer Menu Content --> 
-            <div class="drawer-content">
-              
-              <mwc-list activatable @selected="${this._logSelected}">
-                <mwc-list-item value="Item0" twoline hasMeta>
-                  <span>Item 0</span>
-                  <span slot="secondary">Secondary line</span>
-                  <mwc-icon slot="meta">info</mwc-icon>
-                </mwc-list-item>
-                <mwc-list-item twoline hasMeta>
-                  <span>Item 1</span>
-                  <span slot="secondary">Secondary line</span>
-                  <mwc-icon slot="meta">info</mwc-icon>
-                </mwc-list-item>
-                <li divider padded role="separator"></li>
-                <mwc-list-item graphic="avatar">
-                  <span>Item 2</span>
-                  <mwc-icon slot="graphic">folder</mwc-icon>
-                </mwc-list-item>
-                <mwc-list-item graphic="avatar">
-                  <span>Item 3</span>
-                  <mwc-icon slot="graphic">folder</mwc-icon>
-                </mwc-list-item>
-              </mwc-list>
-              
-              <!-- button to close the drawer for full width drawer --> 
-              <mwc-icon-button
-                icon="close"
-                class="drawer-close-icon"
-                @click="${() => this.drawerIsOpen = false}">
-              </mwc-icon-button>
-
-            </div>
-
-            <!-- Drawer Content --> 
-            <div id="container" slot="appContent">
-              <mwc-top-app-bar @MDCTopAppBar:nav="${this._handleDrawer}">
-                <!-- navigationIcon fire the @MDCTopAppBar:nav itself --> 
-                <mwc-icon-button slot="navigationIcon" icon="menu"></mwc-icon-button>
-                
-                <div slot="title">${this.title}</div>
-
-                <!-- example of handle click event --> 
-                <mwc-icon-button
-                  icon="message"
-                  slot="actionItems"
-                  @click="${() => console.log('@CLICK >> Message')}">
-                </mwc-icon-button>
-
-                <!-- Offline Icon --> 
-                <mwc-icon 
-                  slot="actionItems"
-                  ?hidden="${!this.offline}"
-                  ?aria-hidden="${!this.offline}">cloud_off</mwc-icon>
-
-              </mwc-top-app-bar>
-
-              <mwc-fab
-                id="fab-menu"
-                icon="menu"
-                label="menu"
-                @click="${this._handleDrawer}">
-              </mwc-fab>
-            </div>
-
-          </mwc-drawer>
+      <!-- Drawer layout -->
     `
 
     const tabsLayout = html`
-      <!-- Tabs --> 
-      <mwc-tab-bar @MDCTabBar:activated="${this._tabsRoute}">
-        <mwc-tab label="home"></mwc-tab>
-        <mwc-tab label="login"></mwc-tab>
-      </mwc-tab-bar>
-    `
-    const materialSnackbar = html`
-      <mwc-snackbar>
-         <mwc-icon-button icon="close" slot="dismiss"></mwc-icon-button>
-      </mwc-snackbar>
+      <!-- Tabs layout --> 
     `
 
     return html`
-      <!-- Main -->
-      <main>
 
-        <!-- Progress Bar for Async tasks -->
-        <mwc-linear-progress 
-          indeterminate 
-          .closed="${!this._hasPendingChildren}">
-        </mwc-linear-progress>
+      <!-- Progress Bar for Async tasks TODO -->
+      <h2>
+        Async Tasks > ${this._hasPendingChildren} (#${this._pendingCount})
+        <a href="/home">Home</a>|<a href="/login">Login</a>|<a href="/xxx">XXX</a>|
+      </h2>
 
-        <!-- Layout --> 
-        ${this.drawerMode ? drawerLayout : tabsLayout}
-        ${this._renderMainContent()}
+      <!-- Layout --> 
+      ${this.drawerMode ? drawerLayout : tabsLayout}
+      ${this._renderMainContent()}
 
-      </main>
-
-      <!-- Snackbar -->
-      ${materialSnackbar}
+      <snack-bar></snack-bar>
     `
   }
 }