+++ src/views/view-notfound.js
@@ -5,6 +5,7 @@ class ViewNotfound extends LitElement {
   static get styles () {
     return css`
       :host {
+        display: block;
         background-color: red;
       }
     `