+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 08-07-2021
+### Added
+  - Login view
+
+### Removed
+  - Material Design MWC
+  - All Material Design components (toolbar, drawer ...)
+
+### Changed
+  - Update Vite, aws-amplify updated to latest version
+
 ## [1.0.0] - 21-06-2021
 ### Added
 - Init the Repo [@CICCIOSGAMINO](https://github.com/CICCIOSGAMINO).